Change the way MetricTag are concatenated.
Now the measurements are concatenated using '.'. The measurement is not
sanitized anymore.
To reintroduce it we would need to have a vector of names that we join
when a format function is called. This currently seems overkill.

Also a new member function has been added to remove all the counters
which match a prefix.